pyproject.toml
[tool.poetry]
name = "todos"
version = "0.1.0"
description = ""
authors = ["Khizar3333 <khizarahmad3333@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.110.0"
uvicorn = {extras = ["standard"], version = "^0.29.0"}
sqlmodel = "^0.0.18"
python-dotenv = "^1.0.1"

[tool.poetry.scripts]
dev= "todos.main:start"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
